kafka_pipeline/requirements.txt
boto3
psycopg2-binary
python-dotenv
pytest
pylint
confluent_kafka
